tests/telegram-qr/python/requirements.txt
flask
python-dotenv
qrcode[pil]
requests
